Extended records to store doubles and floats
This Commit stores fields in records that are float and integers as
floats and integers.

+++ src/main/java/controller/ImportController.java
@@ -237,8 +237,8 @@ public class ImportController extends SubController {
 				case "Int":
 					colNames[i].setType(ColumnType.INT);
 					break;
-				case "Float":
-					colNames[i].setType(ColumnType.FLOAT);
+				case "Double":
+					colNames[i].setType(ColumnType.DOUBLE);
 					break;
 				case "Time":
 					colNames[i].setType(ColumnType.TIME);
@@ -260,7 +260,6 @@ public class ImportController extends SubController {
 
 			Group g = new Group(gli.txtField.getText(), gli.box
 					.getSelectionModel().getSelectedItem(), colNames,
-
 					gli.primKey);
 			
 			for (FileListItem fli : gli.fileList) {
@@ -504,7 +503,7 @@ public class ImportController extends SubController {
 				    FXCollections.observableArrayList(
 				        "String",
 				        "Int",
-				        "Float",
+				        "Double",
 				        "Time",
 				        "Date",
 				        "Date/Time",
